Deleting emacs swap file
By the way, there is a mod to the bios to jump past an instruction. The first mod is at line 57602 in which 74 is changed to eb to make a short jump always instead of at zero. The second mod is at line 65537 in which ff is changed to 88 to make sure checksum passes. This has not been verified yet.